Refactoring and initial import of all screens

+++ CorpMediaJson/src/com/intuiture/corp/json/CityCompensatoryAllowanceJson.java
@@ -1,37 +1,42 @@
 package com.intuiture.corp.json;
 
 public class CityCompensatoryAllowanceJson {
-	
+
 	private Integer cityCompensatoryAllowanceId;
-	private Integer maxAnnualLimit;
+	private Long maxAnnualLimit;
 	private Integer requireSubmissionId;
 	private Integer companyId;
-	
+
 	public Integer getCityCompensatoryAllowanceId() {
 		return cityCompensatoryAllowanceId;
 	}
+
 	public void setCityCompensatoryAllowanceId(Integer cityCompensatoryAllowanceId) {
 		this.cityCompensatoryAllowanceId = cityCompensatoryAllowanceId;
 	}
-	public Integer getMaxAnnualLimit() {
+
+	public Long getMaxAnnualLimit() {
 		return maxAnnualLimit;
 	}
-	public void setMaxAnnualLimit(Integer maxAnnualLimit) {
+
+	public void setMaxAnnualLimit(Long maxAnnualLimit) {
 		this.maxAnnualLimit = maxAnnualLimit;
 	}
+
 	public Integer getRequireSubmissionId() {
 		return requireSubmissionId;
 	}
+
 	public void setRequireSubmissionId(Integer requireSubmissionId) {
 		this.requireSubmissionId = requireSubmissionId;
 	}
+
 	public Integer getCompanyId() {
 		return companyId;
 	}
+
 	public void setCompanyId(Integer companyId) {
 		this.companyId = companyId;
 	}
-	
-	
 
 }
