+++ CorpMediaJson/src/com/intuiture/corp/json/GratuityContributionJson.java
@@ -1,31 +1,33 @@
 package com.intuiture.corp.json;
 
 public class GratuityContributionJson {
-	
+
 	private Integer gratuityContributionId;
-	private Integer maxAnnualLimit;
+	private Long maxAnnualLimit;
 	private Integer companyId;
-	
+
 	public Integer getGratuityContributionId() {
 		return gratuityContributionId;
 	}
+
 	public void setGratuityContributionId(Integer gratuityContributionId) {
 		this.gratuityContributionId = gratuityContributionId;
 	}
-	public Integer getMaxAnnualLimit() {
+
+	public Long getMaxAnnualLimit() {
 		return maxAnnualLimit;
 	}
-	public void setMaxAnnualLimit(Integer maxAnnualLimit) {
+
+	public void setMaxAnnualLimit(Long maxAnnualLimit) {
 		this.maxAnnualLimit = maxAnnualLimit;
 	}
+
 	public Integer getCompanyId() {
 		return companyId;
 	}
+
 	public void setCompanyId(Integer companyId) {
 		this.companyId = companyId;
 	}
-	
-	
-	
 
 }
