+++ CorpMediaJson/src/com/intuiture/corp/json/ProfessionalAllowanceJson.java
@@ -1,37 +1,42 @@
 package com.intuiture.corp.json;
 
 public class ProfessionalAllowanceJson {
-	
+
 	private Integer professionalAllowanceId;
-	private Integer maxAnnualLimit;
+	private Long maxAnnualLimit;
 	private Integer requireSubmissionId;
 	private Integer companyId;
-	
+
 	public Integer getProfessionalAllowanceId() {
 		return professionalAllowanceId;
 	}
+
 	public void setProfessionalAllowanceId(Integer professionalAllowanceId) {
 		this.professionalAllowanceId = professionalAllowanceId;
 	}
-	public Integer getMaxAnnualLimit() {
+
+	public Long getMaxAnnualLimit() {
 		return maxAnnualLimit;
 	}
-	public void setMaxAnnualLimit(Integer maxAnnualLimit) {
+
+	public void setMaxAnnualLimit(Long maxAnnualLimit) {
 		this.maxAnnualLimit = maxAnnualLimit;
 	}
+
 	public Integer getRequireSubmissionId() {
 		return requireSubmissionId;
 	}
+
 	public void setRequireSubmissionId(Integer requireSubmissionId) {
 		this.requireSubmissionId = requireSubmissionId;
 	}
+
 	public Integer getCompanyId() {
 		return companyId;
 	}
+
 	public void setCompanyId(Integer companyId) {
 		this.companyId = companyId;
 	}
-	
-	
 
 }
