+++ CorpMediaService/src/com/intuiture/corp/entity/TransportAllowance.java
@@ -14,53 +14,65 @@ public class TransportAllowance {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Integer transportAllowanceId;
-	private Integer maxAnnualLimit;
+	private Long maxAnnualLimit;
 	private Integer requireSubmissionId;
 	private Boolean isDeleted;
 	private Date createdOn;
 	private Date updatedOn;
 	private Integer companyId;
-	
-	
+
 	public Integer getTransportAllowanceId() {
 		return transportAllowanceId;
 	}
+
 	public void setTransportAllowanceId(Integer transportAllowanceId) {
 		this.transportAllowanceId = transportAllowanceId;
 	}
-	public Integer getMaxAnnualLimit() {
+
+	public Long getMaxAnnualLimit() {
 		return maxAnnualLimit;
 	}
-	public void setMaxAnnualLimit(Integer maxAnnualLimit) {
+
+	public void setMaxAnnualLimit(Long maxAnnualLimit) {
 		this.maxAnnualLimit = maxAnnualLimit;
 	}
+
 	public Integer getRequireSubmissionId() {
 		return requireSubmissionId;
 	}
+
 	public void setRequireSubmissionId(Integer requireSubmissionId) {
 		this.requireSubmissionId = requireSubmissionId;
 	}
+
 	public Boolean getIsDeleted() {
 		return isDeleted;
 	}
+
 	public void setIsDeleted(Boolean isDeleted) {
 		this.isDeleted = isDeleted;
 	}
+
 	public Date getCreatedOn() {
 		return createdOn;
 	}
+
 	public void setCreatedOn(Date createdOn) {
 		this.createdOn = createdOn;
 	}
+
 	public Date getUpdatedOn() {
 		return updatedOn;
 	}
+
 	public void setUpdatedOn(Date updatedOn) {
 		this.updatedOn = updatedOn;
 	}
+
 	public Integer getCompanyId() {
 		return companyId;
 	}
+
 	public void setCompanyId(Integer companyId) {
 		this.companyId = companyId;
 	}
